requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
requests
bs4
Unidecode==1.3.7
uWSGI